Create Models from pydantic as restur structures

+++ services/calculator.py
@@ -1,15 +1,15 @@
-def add (a:int, b:int)->int:
+def add (a:int, b:int)->float:
     """ Add two numbers"""
     return a + b
 
-def sub (a:int, b:int)->int:
+def sub (a:int, b:int)->float:
     """ Sub two numbers"""
     return a - b
 
-def mul (a:int, b:int)->int:
+def mul (a:int, b:int)->float:
     """ Mul two numbers"""
     return a * b
 
-def div (a:int, b:int)->int:
+def div (a:int, b:int)->float:
     """ Mul two numbers"""
     return a / b
